+++ githarbor/__init__.py
@@ -1,3 +1,3 @@
 """GitHarbor application package."""
 
-__version__ = "0.6.4"
+__version__ = "0.6.5"